+++ src/ska_sdp_instrumental_calibration/workflow/pipelines/bandpass_calibration.py
@@ -16,8 +16,8 @@
 
 import warnings
 
-import dask.array as da
 import numpy as np
+from astropy.coordinates import SkyCoord
 from dask.distributed import Client, LocalCluster
 from ska_sdp_datamodels.calibration.calibration_functions import (
     export_gaintable_to_hdf5,
@@ -25,15 +25,12 @@ from ska_sdp_datamodels.calibration.calibration_functions import (
 from ska_sdp_func_python.preprocessing.flagger import rfi_flagger
 
 from ska_sdp_instrumental_calibration.data_managers.dask_wrappers import (
-    apply_gaintable_to_dataset,
     load_ms,
     predict_vis,
     run_solver,
 )
 from ska_sdp_instrumental_calibration.logger import setup_logger
-from ska_sdp_instrumental_calibration.processing_tasks.lsm_tmp import (
-    generate_lsm,
-)
+from ska_sdp_instrumental_calibration.processing_tasks.lsm import generate_lsm
 from ska_sdp_instrumental_calibration.workflow.utils import create_demo_ms
 
 warnings.simplefilter(action="ignore", category=FutureWarning)
@@ -51,51 +48,88 @@ def run(pipeline_config) -> None:
         None
     """
 
-    # Required external data
-    gleamfile = pipeline_config.get("gleamfile", None)
-    if gleamfile is None:
-        raise ValueError("GLEAM catalogue gleamegc.dat is required")
-    eb_ms = pipeline_config.get("eb_ms", None)
-    if eb_ms is None:
-        raise ValueError("Name of Everybeam mock Measurement Set is required")
-    eb_coeffs = pipeline_config.get("eb_coeffs", None)
-    if eb_coeffs is None:
-        raise ValueError("Path to Everybeam coeffs directory is required")
-
-    # Filename
+    # Filenames
     ms_name = pipeline_config.get("ms_name", "demo.ms")
     hdf5_name = pipeline_config.get("hdf5_name", "demo.hdf5")
 
     # Sky model info
+    lsm = pipeline_config.get("lsm", None)
     fov = pipeline_config.get("fov_deg", 10)
     flux_limit = pipeline_config.get("flux_limit", 1)
+    gleamfile = pipeline_config.get("gleamfile", None)
+    # Check required external data
+    if lsm is None and gleamfile is None:
+        raise ValueError("Either a LSM or a catalogue file is required")
+
+    # Beam model info
+    beam_type = pipeline_config.get("beam_type", "everybeam")
+    eb_coeffs = pipeline_config.get("eb_coeffs", None)
+    eb_ms = pipeline_config.get("eb_ms", ms_name)
+    if beam_type.lower() == "everybeam":
+        # Required external data
+        if eb_coeffs is None:
+            raise ValueError("Path to Everybeam coeffs directory is required")
+        logger.info(f"Initialising the EveryBeam telescope model with {eb_ms}")
+    elif beam_type.lower() == "none":
+        logger.info("Predicting model visibilities without a beam")
+    else:
+        raise ValueError(f"Unknown beam type: {beam_type}")
 
     # Corruption and solver type
     gains = pipeline_config.get("gains", True)
     leakage = pipeline_config.get("leakage", False)
+    rotation = pipeline_config.get("rotation", False)
     solver = pipeline_config.get("solver", "gain_substitution")
+    refant = 0
 
     if ms_name == "demo.ms":
         # Generate a demo MSv2 Measurement Set
+        ntimes = pipeline_config.get("ntimes", 1)
+        nchannels = pipeline_config.get("nchannels", 64)
+        phasecentre = SkyCoord(ra=0.0, dec=-27.0, unit="degree")
+        if lsm is None:
+            # Get the LSM (single call for all channels / dask tasks)
+            logger.info(f"LSM: {gleamfile} < {fov/2} deg > {flux_limit} Jy")
+            lsm = generate_lsm(
+                gleamfile=gleamfile,
+                phasecentre=phasecentre,
+                fov=fov,
+                flux_limit=flux_limit,
+            )
+            logger.info(f"LSM: found {len(lsm)} components")
+
         logger.info(f"Generating a demo MSv2 Measurement Set {ms_name}")
-        create_demo_ms(
+        truetable = create_demo_ms(
             ms_name=ms_name,
+            ntimes=ntimes,
+            nchannels=nchannels,
             gains=gains,
             leakage=leakage,
-            gleamfile=gleamfile,
-            eb_ms=eb_ms,
+            rotation=rotation,
+            phasecentre=phasecentre,
+            lsm=lsm,
+            beam_type=beam_type,
             eb_coeffs=eb_coeffs,
+            eb_ms=eb_ms,
         )
 
+    # Dask info
+    cluster = pipeline_config.get("dask_cluster", None)
+    # The number of channels per frequency chunk
+    fchunk = pipeline_config.get("fchunk", 16)
+
+    logger.info(f"Starting pipeline with {fchunk}-channel chunks")
+
     # Set up a local dask cluster and client
-    cluster = LocalCluster()
+    if cluster is None:
+        logger.info("No dask cluster supplied. Using LocalCluster")
+        cluster = LocalCluster()
+    else:
+        logger.info("Using existing dask cluster")
     client = Client(cluster)
 
-    # Set the number of channels per frequency chunk
-    fchunk = 16
-
     # Read in the Visibility dataset
-    logger.info(f"Reading {ms_name} in {fchunk}-channel chunks")
+    logger.info(f"Setting input from {ms_name} in {fchunk}-channel chunks")
     vis = load_ms(ms_name, fchunk)
 
     # Pre-processing
@@ -103,53 +137,68 @@ def run(pipeline_config) -> None:
     #  - Move to dask_wrappers? RFI flagging may need bandwidth...
     rfi_flagging = False
     if rfi_flagging:
-        logger.info("Calling ska-sdp-func RFI flagger")
+        logger.info("Setting the ska-sdp-func RFI flagger")
         vis = rfi_flagger(vis)
 
-    # Get the LSM (single call for all channels / dask tasks)
-    logger.info(f"Generating {gleamfile} LSM < {fov/2} deg > {flux_limit} Jy")
-    lsm = generate_lsm(
-        gleamfile=gleamfile,
-        phasecentre=vis.phasecentre,
-        fov=fov,
-        flux_limit=flux_limit,
-    )
+    # Set up the local sky model (single call for all channels)
+    #  - Could do this earlier, but currently use vis for phase centre
+    if lsm is None:
+        logger.info(f"LSM: {gleamfile} < {fov/2} deg > {flux_limit} Jy")
+        lsm = generate_lsm(
+            gleamfile=gleamfile,
+            phasecentre=vis.phasecentre,
+            fov=fov,
+            flux_limit=flux_limit,
+        )
+        logger.info(f"LSM: found {len(lsm)} components")
 
     # Predict model visibilities
-    logger.info(f"Predicting model visibilities in {fchunk}-channel chunks")
-    modelvis = predict_vis(vis, lsm, eb_ms=eb_ms, eb_coeffs=eb_coeffs)
-
-    # Call the solver
-    logger.info(f"Running calibration in {fchunk}-channel chunks")
-    gaintable = run_solver(
-        vis=vis, modelvis=modelvis, solver=solver, niter=50, refant=0
+    logger.info(f"Setting vis predict in {fchunk}-channel chunks")
+    modelvis = predict_vis(
+        vis, lsm, beam_type=beam_type, eb_ms=eb_ms, eb_coeffs=eb_coeffs
     )
 
+    # Call the solver
+    logger.info(f"Setting calibration in {fchunk}-channel chunks")
+    gaintable = run_solver(
+        vis=vis, modelvis=modelvis, solver=solver, niter=50, refant=refant
+    )
+
+    # Output hdf5 file
+    logger.info("Running graph and returning calibration solutions")
+    gaintable.load()
+    logger.info(f"Writing solutions to {hdf5_name}")
+    export_gaintable_to_hdf5([gaintable], hdf5_name)
+
     # Convergence checks (noise-free demo version)
-    #  - Note that this runs the graph again. I tried assigning both vis
-    #    and modelvis to gaintable for a single load, but it got confused
-    #    by the baseline MultiIndex. MultiIndex causes a lot of trouble...
-    #  - This is just a quick check, so it shouldn't hurt to run it again.
     if ms_name == "demo.ms":
-        logger.info("Applying solutions")
-        vis = apply_gaintable_to_dataset(vis, gaintable, inverse=True)
         logger.info("Checking results")
+        gfit = gaintable.gain.data
+        true = truetable.gain.data
+        # Reference all polarisations again the X gain for the ref antenna
+        gfit *= np.exp(
+            -1j
+            * np.angle(gfit[:, [refant], :, 0, 0][..., np.newaxis, np.newaxis])
+        )
+        true *= np.exp(
+            -1j
+            * np.angle(true[:, [refant], :, 0, 0][..., np.newaxis, np.newaxis])
+        )
         if solver == "gain_substitution":
-            # Don't expect convergence for these, so zeros them
-            vis.vis.data[..., 1] = da.zeros(vis.vis.shape[:3], "complex")
-            vis.vis.data[..., 2] = da.zeros(vis.vis.shape[:3], "complex")
-            modelvis.vis.data[..., 1] = da.zeros(vis.vis.shape[:3], "complex")
-            modelvis.vis.data[..., 2] = da.zeros(vis.vis.shape[:3], "complex")
-        converged = np.allclose(modelvis.vis.data, vis.vis.data, atol=1e-6)
+            # For independent Y calibration, reference Y gains separately
+            gfit[:, :, :, 1, 1] *= np.exp(
+                -1j * np.angle(gfit[:, [refant], :, 1, 1])
+            )
+            true[:, :, :, 1, 1] *= np.exp(
+                -1j * np.angle(true[:, [refant], :, 1, 1])
+            )
+            # Fit won't change off-diag, so zero those before comparisons
+            true[..., 0, 1] = true[..., 1, 0] = 0
+        converged = np.allclose(gfit, true, atol=1e-6)
         if converged:
             logger.info("Convergence checks passed")
         else:
             logger.warning("Solving failed")
-
-    # Output hdf5 file
-    logger.info(f"Writing solutions to {hdf5_name}")
-    gaintable.load()
-    export_gaintable_to_hdf5([gaintable], hdf5_name)
 
     # Shut down the scheduler and workers
     client.close()