+++ src/ska_sdp_instrumental_calibration/workflow/pipelines/bandpass_polarisation.py
@@ -17,6 +17,7 @@
 import warnings
 
 import numpy as np
+from astropy.coordinates import SkyCoord
 from dask.distributed import Client, LocalCluster
 from ska_sdp_datamodels.calibration.calibration_functions import (
     export_gaintable_to_hdf5,
@@ -28,15 +29,12 @@ from ska_sdp_func_python.preprocessing.averaging import (
 from ska_sdp_func_python.preprocessing.flagger import rfi_flagger
 
 from ska_sdp_instrumental_calibration.data_managers.dask_wrappers import (
-    apply_gaintable_to_dataset,
     load_ms,
     predict_vis,
     run_solver,
 )
 from ska_sdp_instrumental_calibration.logger import setup_logger
-from ska_sdp_instrumental_calibration.processing_tasks.lsm_tmp import (
-    generate_lsm,
-)
+from ska_sdp_instrumental_calibration.processing_tasks.lsm import generate_lsm
 from ska_sdp_instrumental_calibration.processing_tasks.post_processing import (
     model_rotations,
 )
@@ -57,24 +55,32 @@ def run(pipeline_config) -> None:
         None
     """
 
-    # Required external data
-    gleamfile = pipeline_config.get("gleamfile", None)
-    if gleamfile is None:
-        raise ValueError("GLEAM catalogue gleamegc.dat is required")
-    eb_ms = pipeline_config.get("eb_ms", None)
-    if eb_ms is None:
-        raise ValueError("Name of Everybeam mock Measurement Set is required")
-    eb_coeffs = pipeline_config.get("eb_coeffs", None)
-    if eb_coeffs is None:
-        raise ValueError("Path to Everybeam coeffs directory is required")
-
-    # Filename
+    # Filenames
     ms_name = pipeline_config.get("ms_name", "demo.ms")
     hdf5_name = pipeline_config.get("hdf5_name", "demo.hdf5")
 
     # Sky model info
+    lsm = pipeline_config.get("lsm", None)
     fov = pipeline_config.get("fov_deg", 10)
     flux_limit = pipeline_config.get("flux_limit", 1)
+    gleamfile = pipeline_config.get("gleamfile", None)
+    # Check required external data
+    if lsm is None and gleamfile is None:
+        raise ValueError("Either a LSM or a catalogue file is required")
+
+    # Beam model info
+    beam_type = pipeline_config.get("beam_type", "everybeam")
+    eb_coeffs = pipeline_config.get("eb_coeffs", None)
+    eb_ms = pipeline_config.get("eb_ms", ms_name)
+    if beam_type.lower() == "everybeam":
+        # Required external data
+        if eb_coeffs is None:
+            raise ValueError("Path to Everybeam coeffs directory is required")
+        logger.info(f"Initialising the EveryBeam telescope model with {eb_ms}")
+    elif beam_type.lower() == "none":
+        logger.info("Predicting model visibilities without a beam")
+    else:
+        raise ValueError(f"Unknown beam type: {beam_type}")
 
     # Pre-processing
     rfi_flagging = False  # not yet set up for dask chunks
@@ -83,27 +89,53 @@ def run(pipeline_config) -> None:
 
     if ms_name == "demo.ms":
         # Generate a demo MSv2 Measurement Set
+        ntimes = pipeline_config.get("ntimes", 1)
+        nchannels = pipeline_config.get("nchannels", 64)
+        phasecentre = SkyCoord(ra=0.0, dec=-27.0, unit="degree")
+        if lsm is None:
+            # Get the LSM (single call for all channels / dask tasks)
+            logger.info(f"LSM: {gleamfile} < {fov/2} deg > {flux_limit} Jy")
+            lsm = generate_lsm(
+                gleamfile=gleamfile,
+                phasecentre=phasecentre,
+                fov=fov,
+                flux_limit=flux_limit,
+            )
+            logger.info(f"LSM: found {len(lsm)} components")
+
         logger.info(f"Generating a demo MSv2 Measurement Set {ms_name}")
-        create_demo_ms(
+        truetable = create_demo_ms(
             ms_name=ms_name,
+            ntimes=ntimes,
+            nchannels=nchannels,
+            wide_channels=True,
             gains=True,
             leakage=True,
             rotation=True,
-            wide_channels=True,
-            gleamfile=gleamfile,
-            eb_ms=eb_ms,
+            phasecentre=phasecentre,
+            lsm=lsm,
+            beam_type=beam_type,
             eb_coeffs=eb_coeffs,
+            eb_ms=eb_ms,
         )
 
+    # Dask info
+    cluster = pipeline_config.get("dask_cluster", None)
+    # The number of channels per frequency chunk
+    fchunk = pipeline_config.get("fchunk", 16)
+
+    logger.info(f"Starting pipeline with {fchunk}-channel chunks")
+
     # Set up a local dask cluster and client
-    cluster = LocalCluster()
+    if cluster is None:
+        logger.info("No dask cluster supplied. Using LocalCluster")
+        cluster = LocalCluster()
+    else:
+        logger.info("Using existing dask cluster")
     client = Client(cluster)
 
-    # Set the number of channels per frequency chunk
-    fchunk = 16
-
     # Read in the Visibility dataset
-    logger.info(f"Reading {ms_name} in {fchunk}-channel chunks")
+    logger.info(f"Setting input from {ms_name} in {fchunk}-channel chunks")
     vis = load_ms(ms_name, fchunk)
 
     # Pre-processing
@@ -127,51 +159,66 @@ def run(pipeline_config) -> None:
         vis = averaging_frequency(vis, freqstep=preproc_ave_frequency)
 
     # Get the LSM (single call for all channels)
-    logger.info(f"Generating {gleamfile} LSM < {fov/2} deg > {flux_limit} Jy")
-    lsm = generate_lsm(
-        gleamfile=gleamfile,
-        phasecentre=vis.phasecentre,
-        fov=fov,
-        flux_limit=flux_limit,
-    )
+    #  - Could do this earlier, but currently use vis for phase centre
+    if lsm is None:
+        logger.info(f"LSM: {gleamfile} < {fov/2} deg > {flux_limit} Jy")
+        lsm = generate_lsm(
+            gleamfile=gleamfile,
+            phasecentre=vis.phasecentre,
+            fov=fov,
+            flux_limit=flux_limit,
+        )
+        logger.info(f"LSM: found {len(lsm)} components")
 
     # Predict model visibilities
-    logger.info(f"Predicting model visibilities in {fchunk}-channel chunks")
-    modelvis = predict_vis(vis, lsm, eb_ms=eb_ms, eb_coeffs=eb_coeffs)
+    logger.info(f"Setting vis predict in {fchunk}-channel chunks")
+    modelvis = predict_vis(
+        vis, lsm, beam_type=beam_type, eb_ms=eb_ms, eb_coeffs=eb_coeffs
+    )
 
     # Call the solver
-    logger.info(f"Running calibration in {fchunk}-channel chunks")
+    logger.info(f"Setting calibration in {fchunk}-channel chunks")
+    refant = 0
     initialtable = run_solver(
         vis=vis,
         modelvis=modelvis,
         solver="jones_substitution",
         niter=20,
-        refant=0,
+        refant=refant,
     )
 
-    # Load all of the solutions into a numpy array and fit for any
-    # differential rotations (single call for all channels).
+    # Load all of the solutions into a numpy array
+    logger.info("Running graph and returning calibration solutions")
+    initialtable.load()
+
+    # Fit for any differential rotations (single call for all channels).
     # Return gaintable filled with pure rotations.
     #  - If the vis data and model can fit into memory, it would be a good
     #    time to make them persistent. Otherwise they will be re-loaded and
     #    re-predicted in the next graph below.
     #  - Alternatively, export them to disk in a chunked way (e.g. to zarr)
     #  - Alternatively, can regenerate. Do this for now.
-    initialtable.load()
+    logger.info("Fitting differential rotations")
     gaintable = model_rotations(initialtable, plot_sample=True).chunk(
         {"frequency": fchunk}
     )
 
     # Call the solver with updated initial solutions
-    logger.info(f"Rerunning calibration in {fchunk}-channel chunks")
+    logger.info(f"Resetting calibration in {fchunk}-channel chunks")
     gaintable = run_solver(
         vis=vis,
         modelvis=modelvis,
         gaintable=gaintable,
         solver="normal_equations",
         niter=50,
-        refant=0,
+        refant=refant,
     )
+
+    # Output hdf5 file
+    logger.info("Running graph and returning calibration solutions")
+    gaintable.load()
+    logger.info(f"Writing solutions to {hdf5_name}")
+    export_gaintable_to_hdf5([gaintable], hdf5_name)
 
     # Convergence checks (noise-free demo version)
     #  - Note that this runs the graph again. I tried assigning both vis
@@ -179,19 +226,23 @@ def run(pipeline_config) -> None:
     #    by the baseline MultiIndex. MultiIndex causes a lot of trouble...
     #  - This is just a quick check, so it shouldn't hurt to run it again.
     if ms_name == "demo.ms":
-        logger.info("Applying solutions")
-        vis = apply_gaintable_to_dataset(vis, gaintable, inverse=True)
         logger.info("Checking results")
-        converged = np.allclose(modelvis.vis.data, vis.vis.data, atol=1e-6)
+        gfit = gaintable.gain.data
+        true = truetable.gain.data
+        # Reference all polarisations again the X gain for the ref antenna
+        gfit *= np.exp(
+            -1j
+            * np.angle(gfit[:, [refant], :, 0, 0][..., np.newaxis, np.newaxis])
+        )
+        true *= np.exp(
+            -1j
+            * np.angle(true[:, [refant], :, 0, 0][..., np.newaxis, np.newaxis])
+        )
+        converged = np.allclose(gfit, true, atol=1e-6)
         if converged:
             logger.info("Convergence checks passed")
         else:
             logger.warning("Solving failed")
-
-    # Output hdf5 file
-    logger.info(f"Writing solutions to {hdf5_name}")
-    gaintable.load()
-    export_gaintable_to_hdf5([gaintable], hdf5_name)
 
     # Shut down the scheduler and workers
     client.close()