+++ tests/unit/test_dask_wrappers.py
@@ -35,7 +35,7 @@ from ska_sdp_instrumental_calibration.data_managers.dask_wrappers import (
 from ska_sdp_instrumental_calibration.processing_tasks.calibration import (
     apply_gaintable,
 )
-from ska_sdp_instrumental_calibration.processing_tasks.lsm_tmp import (
+from ska_sdp_instrumental_calibration.processing_tasks.lsm import (
     Component,
     convert_model_to_skycomponents,
 )